Support configuration directories.
If the initial configuration file name, or a parameter to require/include,
is a directory, read its content, and for each entry there which ends
in ".conf", recursively read that file (or directory). The directory
entries are not processed in any specific order.

The new test case require_dir_basic shows off the functionality,
including recursively reading a subdirectory of the first level directory.