BROKEN. Work on '.' concat operator and constant replacement.
For string definitions and include/require paths, a '.' concat operator
can be used to build the strings from pieces. This works on the syntax
side.

Also, to make that really usefil, unquoted values which happen to be
already defined constants, should be replaced by their constant value.
That part is BROKEN in this commit, the PHP internal stuff in
replace_constants(), does not work as intended.